Syncing, done.

+++ src/main/java/snownee/cuisine/cookware/client/OvenScreen.java
@@ -45,7 +45,7 @@ public class OvenScreen extends ContainerScreen<OvenContainer> {
     @Override
     protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY) {
         this.font.drawString(this.title.getFormattedText(), 28.0F, 6.0F, 4210752);
-        this.font.drawString(this.playerInventory.getDisplayName().getFormattedText(), 8.0F, (float) (this.ySize - 96 + 2), 4210752);
+        this.font.drawString(this.playerInventory.getDisplayName().getFormattedText(), 8.0F, this.ySize - 96 + 2, 4210752);
     }
 
 }
